KooCore-D/requirements.txt
# =============================================================================
# Momentum Trading System - Dependencies
# =============================================================================

# Core dependencies
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.24.0
lxml>=4.9.0
html5lib>=1.1
requests>=2.28.0

# LLM Integration
openai>=1.0.0
anthropic>=0.25.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Type validation (optional but recommended)
pydantic>=2.0.0

# API Server (optional - for dashboard)
fastapi>=0.100.0
uvicorn>=0.22.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
# mypy>=1.0.0

# Security scanning (optional)
# safety>=2.3.0
# bandit>=1.7.0

# Redis support (optional - for distributed caching)
# redis>=4.5.0
